move tso to pkg
