+++ tests/server/tso/tso_test.go
@@ -24,10 +24,10 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/stretchr/testify/require"
+	"github.com/tikv/pd/pkg/tso"
 	"github.com/tikv/pd/pkg/utils/grpcutil"
 	"github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/server/config"
-	"github.com/tikv/pd/server/tso"
 	"github.com/tikv/pd/tests"
 )
 
@@ -53,7 +53,7 @@ func TestLoadTimestamp(t *testing.T) {
 
 	lastTSMap := requestLocalTSOs(re, cluster, dcLocationConfig)
 
-	re.NoError(failpoint.Enable("github.com/tikv/pd/server/tso/systemTimeSlow", `return(true)`))
+	re.NoError(failpoint.Enable("github.com/tikv/pd/pkg/tso/systemTimeSlow", `return(true)`))
 
 	// Reboot the cluster.
 	re.NoError(cluster.StopAll())
@@ -70,7 +70,7 @@ func TestLoadTimestamp(t *testing.T) {
 		re.Greater(newTS.GetPhysical()-lastTS.GetPhysical(), int64(0))
 	}
 
-	failpoint.Disable("github.com/tikv/pd/server/tso/systemTimeSlow")
+	failpoint.Disable("github.com/tikv/pd/pkg/tso/systemTimeSlow")
 }
 
 func requestLocalTSOs(re *require.Assertions, cluster *tests.TestCluster, dcLocationConfig map[string]string) map[string]*pdpb.Timestamp {
